# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação de API de reservas de hoteis",
        "description": "Esta é a documentação da API para gerenciamento de reservas de hotel.",
        "contact": {
            "name": "Rodrigo Costa",
            "email": "rodrigo321.costa@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reserves": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Cria uma nova reserva",
                "operationId": "23dda9f8e072d7fdd6228cb004b0f484",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "room_id": {
                                        "description": "ID do quarto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "check_in": {
                                        "description": "Data de check-in",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-11-01"
                                    },
                                    "check_out": {
                                        "description": "Data de check-out",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-11-05"
                                    },
                                    "total": {
                                        "description": "Valor total da reserva",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID da reserva",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "room_id": {
                                            "description": "ID do quarto",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "check_in": {
                                            "description": "Data de check-in",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-01"
                                        },
                                        "check_out": {
                                            "description": "Data de check-out",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-05"
                                        },
                                        "total": {
                                            "description": "Valor total da reserva",
                                            "type": "number",
                                            "format": "float",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reserves/{id}/guests": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Adiciona um hóspede a uma reserva",
                "operationId": "03b67e94f5a236861dd89e6fb85e2d38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "document"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do hóspede",
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "document": {
                                        "description": "Documento do hóspede",
                                        "type": "string",
                                        "example": "12345678900"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hóspede adicionado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do hóspede",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome do hóspede",
                                            "type": "string",
                                            "example": "João Silva"
                                        },
                                        "document": {
                                            "description": "Documento do hóspede",
                                            "type": "string",
                                            "example": "12345678900"
                                        },
                                        "reserve_id": {
                                            "description": "ID da reserva associada",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva não registrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Reserva não registrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reserves/{id}/payments": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Adiciona um pagamento a uma reserva",
                "operationId": "f67de5245464d7cb34b419b8fa92b6b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "description": "Valor do pagamento",
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pagamento adicionado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do pagamento",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "value": {
                                            "description": "Valor do pagamento",
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "reserve_id": {
                                            "description": "ID da reserva associada",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva não registrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Reserva não registrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "O valor do pagamento não pode ser maior que o total da reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "O valor do pagamento não pode ser maior que o total da reserva."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}